add animation effect

+++ app/src/main/java/com/example/testapi/base/BaseViewHolder.kt
@@ -5,5 +5,5 @@ import androidx.recyclerview.widget.RecyclerView
 
 
 abstract class BaseViewHolder<T>(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    abstract fun bind(item: T, callback: (() -> Unit)? = null)
+    abstract fun bind(item: T, callback: ((Int) -> Unit)? = null)
 }
